Development: Move the Organization feature out of core into account
Relocate the Organization entity, DTOs, repositories, specs, service and REST
resource from core into the account module, where organization membership is
semantically anchored (it is driven by email-pattern indexing over account.User,
and the entire query layer joins on User):
- core/domain/Organization, core/dto/Organization{DTO,CourseDTO,MemberDTO} -> account
- core/repository/{Organization,CustomOrganization}Repository(+Impl), OrganizationSpecs -> account/repository
- core/service/OrganizationService, core/web/OrganizationResource -> account

OrganizationResource now uses the account dual mapping
{ "api/account/", AccountLegacyRestPaths.CORE_PREFIX }, preserving the existing
api/core/organizations/... URLs (zero client change) while adopting the
module-name prefix. The Organization static metamodel (Organization_) moves with
the entity. Entity-usage arch baselines move with the resource (core return 1->0,
account return 0->1); AdminOrganizationResource stays in admin, UserDTO stays in
core (still referencing the entity via account.domain.Organization).
