+++ src/main/java/de/tum/cit/aet/artemis/account/dto/OrganizationDTO.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.artemis.core.dto;
+package de.tum.cit.aet.artemis.account.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
