+++ src/main/java/de/tum/cit/aet/artemis/core/service/EnrollmentService.java
@@ -12,8 +12,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.lang.CheckReturnValue;
 import org.springframework.stereotype.Service;
 
+import de.tum.cit.aet.artemis.account.domain.Organization;
 import de.tum.cit.aet.artemis.account.domain.User;
-import de.tum.cit.aet.artemis.core.domain.Organization;
 import de.tum.cit.aet.artemis.core.exception.AccessForbiddenException;
 import de.tum.cit.aet.artemis.course.domain.Course;
 import de.tum.cit.aet.artemis.course.repository.CourseRepository;
